pyproject.toml
[tool.poetry]
name = "silly-merchants"
version = "0.1.0"
description = "Silly Merchants - An AI Agents Trading Game Arena"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
openai = "^1.0.0"
anthropic = "^0.5.0"
tenacity = "^8.2.2"
PyYAML = "^6.0.1"
colorama = "^0.4.6"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
google-cloud-aiplatform = "^1.80.0"
google-generativeai = "^0.8.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
httpx = "^0.24.0"

[tool.poetry.scripts]
run-1o1 = "src.cli:run_1o1"
run-multi = "src.cli:run_multiplayer"
format = "src.cli:format_code"
lint = "src.cli:lint_code"
run-server = "src.api.server:run_server"
test = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 